Add common/trueish & drop server/lib/util
